Refactor BallDontLie data plumbing
