+++ README.md
@@ -69,7 +69,7 @@ Runtime source: /public/data/rosters.json generated by scripts/fetch_rosters.ts 
 The players.html page must render from /public/data/rosters.json and never ship season-locked rosters.
 rhicksrad.github.io
 
-Set `BDL_API_KEY` (aliases: `BALLDONTLIE_API_KEY`, `BALL_DONT_LIE_API_KEY`) to your Ball Don't Lie All-Star key before running any fetch or build scripts. The repository no longer includes a default key.
+Set `BDL_API_KEY` (aliases: `BALLDONTLIE_API_KEY`, `BALL_DONT_LIE_API_KEY`) to your Ball Don't Lie All-Star key before running any fetch or build scripts. The repository no longer includes a default key. You can also drop the key in `secrets/bdl_api_key` (or `bdl_api_key.txt`), which the data pipeline now reads automatically if no environment variable is present.
 
 A scheduled job refreshes rosters and deploys if content hashes change.
